zcs-8835:rename ResumeDeviceRequest
rename ResumeDeviceRequest to AllowDeviceRequest, rename all constants and implementation.
