+++ common/src/java/com/zimbra/common/soap/SyncAdminConstants.java
@@ -41,8 +41,8 @@ public final class SyncAdminConstants {
     public static final String E_SUSPEND_DEVICE_RESPONSE = "SuspendDeviceResponse";
     public static final String E_BLOCK_DEVICE_REQUEST = "BlockDeviceRequest";
     public static final String E_BLOCK_DEVICE_RESPONSE = "BlockDeviceResponse";
-    public static final String E_RESUME_DEVICE_REQUEST = "ResumeDeviceRequest";
-    public static final String E_RESUME_DEVICE_RESPONSE = "ResumeDeviceResponse";
+    public static final String E_ALLOW_DEVICE_REQUEST = "AllowDeviceRequest";
+    public static final String E_ALLOW_DEVICE_RESPONSE = "AllowDeviceResponse";
     public static final String E_GET_SYNC_STATE_REQUEST = "GetSyncStateRequest";
     public static final String E_GET_SYNC_STATE_RESPONSE = "GetSyncStateResponse";
     public static final String E_REMOVE_STALE_DEVICE_METADATA_REQUEST = "RemoveStaleDeviceMetadataRequest";
@@ -66,8 +66,8 @@ public final class SyncAdminConstants {
     public static final QName SUSPEND_DEVICE_RESPONSE = QName.get(E_SUSPEND_DEVICE_RESPONSE, NAMESPACE);
     public static final QName BLOCK_DEVICE_REQUEST = QName.get(E_BLOCK_DEVICE_REQUEST, NAMESPACE);
     public static final QName BLOCK_DEVICE_RESPONSE = QName.get(E_BLOCK_DEVICE_RESPONSE, NAMESPACE);
-    public static final QName RESUME_DEVICE_REQUEST = QName.get(E_RESUME_DEVICE_REQUEST, NAMESPACE);
-    public static final QName RESUME_DEVICE_RESPONSE = QName.get(E_RESUME_DEVICE_RESPONSE, NAMESPACE);
+    public static final QName ALLOW_DEVICE_REQUEST = QName.get(E_ALLOW_DEVICE_REQUEST, NAMESPACE);
+    public static final QName ALLOW_DEVICE_RESPONSE = QName.get(E_ALLOW_DEVICE_RESPONSE, NAMESPACE);
     public static final QName GET_SYNC_STATE_REQUEST = QName.get(E_GET_SYNC_STATE_REQUEST, NAMESPACE);
     public static final QName GET_SYNC_STATE_RESPONSE = QName.get(E_GET_SYNC_STATE_RESPONSE, NAMESPACE);
     public static final QName REMOVE_STALE_DEVICE_METADATA_REQUEST = QName.get(E_REMOVE_STALE_DEVICE_METADATA_REQUEST, NAMESPACE);
