+++ soap/src/java/com/zimbra/soap/admin/message/AllowDeviceRequest.java
@@ -1,7 +1,7 @@
 /*
  * ***** BEGIN LICENSE BLOCK *****
  * Zimbra Collaboration Suite Server
- * Copyright (C) 2012, 2013, 2014, 2016 Synacor, Inc.
+ * Copyright (C) 2012, 2013, 2014, 2016, 2020 Synacor, Inc.
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software Foundation,
@@ -36,30 +36,30 @@ import com.zimbra.soap.type.AccountSelector;
  * This will cause a policy reset, but will not reset sync data.
  */
 @XmlAccessorType(XmlAccessType.NONE)
-@XmlRootElement(name=SyncAdminConstants.E_RESUME_DEVICE_REQUEST)
-public class ResumeDeviceRequest {
+@XmlRootElement(name=SyncAdminConstants.E_ALLOW_DEVICE_REQUEST)
+public class AllowDeviceRequest {
 
     /**
      * @zm-api-field-description Account selector
      */
-    @XmlElement(name=AdminConstants.E_ACCOUNT, required=true)
+    @XmlElement(name=AdminConstants.E_ACCOUNT, required=false)
     private AccountSelector account;
 
     /**
      * @zm-api-field-tag device-id
      * @zm-api-field-description Device ID
      */
-    @XmlElement(name=SyncConstants.E_DEVICE, required=false)
+    @XmlElement(name=SyncConstants.E_DEVICE, required=true)
     private DeviceId deviceId;
 
     /**
      * no-argument constructor wanted by JAXB
      */
     @SuppressWarnings("unused")
-    private ResumeDeviceRequest() {
+    private AllowDeviceRequest() {
     }
 
-    public ResumeDeviceRequest(AccountSelector account) {
+    public AllowDeviceRequest(AccountSelector account) {
         this.account = account;
     }
