+++ soap/src/java/com/zimbra/soap/admin/message/AllowDeviceResponse.java
@@ -1,7 +1,7 @@
 /*
  * ***** BEGIN LICENSE BLOCK *****
  * Zimbra Collaboration Suite Server
- * Copyright (C) 2011, 2012, 2013, 2014, 2016 Synacor, Inc.
+ * Copyright (C) 2012, 2013, 2014, 2016, 2020 Synacor, Inc.
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software Foundation,
@@ -15,35 +15,55 @@
  * ***** END LICENSE BLOCK *****
  */
 
-package com.zimbra.soap.sync.message;
+package com.zimbra.soap.admin.message;
 
-import com.google.common.base.MoreObjects;
+import java.util.Collections;
+import java.util.List;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlType;
 
+import com.google.common.base.MoreObjects;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
+import com.zimbra.common.soap.SyncAdminConstants;
 import com.zimbra.common.soap.SyncConstants;
-import com.zimbra.soap.sync.type.DeviceStatusInfo;
+import com.zimbra.soap.admin.type.DeviceStatusInfo;
 
 @XmlAccessorType(XmlAccessType.NONE)
-@XmlRootElement(name=SyncConstants.E_RESUME_DEVICE_RESPONSE)
-public class ResumeDeviceResponse {
+@XmlRootElement(name=SyncAdminConstants.E_ALLOW_DEVICE_RESPONSE)
+@XmlType(propOrder = {})
+public class AllowDeviceResponse {
 
     /**
-     * @zm-api-field-description Device status information
+     * @zm-api-field-description Information about device status
      */
     @XmlElement(name=SyncConstants.E_DEVICE /* device */, required=false)
-    private DeviceStatusInfo device;
+    private List<DeviceStatusInfo> devices = Lists.newArrayList();
 
-    public ResumeDeviceResponse() {
+    public AllowDeviceResponse() {
     }
 
-    public void setDevice(DeviceStatusInfo device) { this.device = device; }
-    public DeviceStatusInfo getDevice() { return device; }
+    public void setDevices(Iterable<DeviceStatusInfo> devices) {
+        this.devices.clear();
+        if (devices != null) {
+            Iterables.addAll(this.devices, devices);
+        }
+    }
+
+    public void addDevice(DeviceStatusInfo device) {
+        this.devices.add(device);
+    }
+
+    public List<DeviceStatusInfo> getDevices() {
+        return Collections.unmodifiableList(devices);
+    }
 
     public MoreObjects.ToStringHelper addToStringInfo(MoreObjects.ToStringHelper helper) {
-        return helper.add("device", device);
+        return helper.add("devices", devices);
     }
 
     @Override
